requirements.txt
fastapi==0.100.0
uvicorn==0.23.0
python-dotenv==1.0.1
python-multipart==0.0.20
sqlalchemy==2.0.37
sqlalchemy-orm
SQLAlchemy==2.0.37
PyJWT==2.10.1
pillow==11.1.0
pyotp==2.9.0
fastapi-cache2==0.2.2
schedule==1.2.2
email_validator==2.2.0
python-jose==3.4.0
psycopg2-binary==2.9.10
passlib==1.7.4
